Fixed #12024: Changed admin code to avoid raising an exception when a field listed in raw_id_fields has limit_choices_to specified as a Q object.
Tweaked a test to trigger the condition and verify the fix. 

Finally, documented that limit_choices_to specified as a Q object has no effect 
on the choices available for fields listed in raw_id_fields, and removed another 
incorrect note that claimed limit_choices_to had no effect on inlines in the admin.



git-svn-id: https://code.djangoproject.com/svn/django/trunk@12728 bcc190cf-cafb-0310-a4f2-bffc1f526a37